REFACTOR: move playlists and soundboards to new files now earch one have you own file.rs for better management of the code and readability
